Finished "Planning & First Steps", Implemented Test Server
Followed along w/ video 2 in the "Building a RESTful API w/ Node.js"
playlist on YouTube. Created a server using Node.js as the runtime
environment and Exress as the web application framework. Was able to
connect to the server w/ localhost:3000 and send RESTful operations
(HTTP GET) w/ Postman.