Implemented password authentication w/ frontend & backend
Followed similar format to v1 of the IMA and added a password
requirement to creating new users. Also included the regex check for
usernames, so that new users must enter properly formatted email
addresses. The frontend reflects the change in the 'Create User' tab and
also checks for and creates an error message for duplicate users.